+++ src/api/definitions.ts
@@ -10,9 +10,13 @@ import {
 import { CurrencyCode } from '../types/currencyCode';
 import { PaymentResult } from '../types/payments';
 import { DataSignature } from '../types/signature';
-import { OrdinalItem } from '../types/items';
+import { CreateItemsOrder, ItemTransferResult, OrdinalItem } from '../types/items';
 
-type PathWithVariable<Prefix extends string, Variable extends string> = `${Prefix}/${Variable}`;
+type PathWithVariable<
+	Prefix extends string,
+	Variable extends string,
+	Suffix extends string
+> = `${Prefix}/${Variable}${Suffix}`;
 
 type Items<T> = {
 	items: T[];
@@ -34,7 +38,7 @@ export type CloudResponse = {
 	'/v1/connect/wallet/pay': PaymentResult;
 	'/v1/connect/wallet/payment': PaymentResult;
 
-	'/v1/connect/runExtension/purse/pay': string;
+	'/v1/connect/runExtension/purse/pay': { partiallySignedTx: string };
 	'/v1/connect/runExtension/purse/broadcast': void;
 	'/v1/connect/runExtension/owner/next': { ownerAddress: string };
 	'/v1/connect/runExtension/owner/sign': { signedTransaction: string };
@@ -42,9 +46,19 @@ export type CloudResponse = {
 
 	'/v3/wallet/items/inventory': Items<OrdinalItem>;
 	'/v3/itemListing/list': Items<OrdinalItem>;
-	'/v3/wallet/items/send': OrdinalItem;
+	'/v3/wallet/items/send': ItemTransferResult;
+	'/v3/wallet/transactions/send/paymentRequest': PaymentResult;
+
+	'/v3/itemCreationOrder': CreateItemsOrder;
+	'/v3/itemCreationOrder/createBatch': CreateItemsOrder;
 } & {
-	[K in PathWithVariable<'/v1/connect/wallet/exchangeRate', CurrencyCode>]: ExchangeRate;
+	[K in PathWithVariable<'/v1/connect/wallet/exchangeRate', CurrencyCode, ''>]: ExchangeRate;
+} & {
+	[K in PathWithVariable<'/v3/itemCreationOrder', string, '/commit'>]: CreateItemsOrder;
+} & {
+	[K in PathWithVariable<'/v3/itemCreationOrder', string, '/add'>]: CreateItemsOrder;
+} & {
+	[K in PathWithVariable<'/v3/itemCreationOrder', string, ''>]: CreateItemsOrder;
 };
 
 export type CloudEndpoint = keyof CloudResponse;